% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopViz.R
\name{.PlotAxes}
\alias{.PlotAxes}
\title{Plot the axes, ticks and labels}
\usage{
.PlotAxes(
  plot,
  person_count,
  reference_person_count,
  treatment_person_count,
  person_spacing,
  reference_name,
  treatment_name,
  treatment_probability,
  reference_probability,
  treatment_confidence_interval
)
}
\arguments{
\item{plot}{{ggplot2} object to which to add elements}

\item{person_count}{Number of people being rendered}

\item{reference_person_count}{Number of people affected by the reference}

\item{treatment_person_count}{number of people affected by the treatment}

\item{person_spacing}{Spacing between people}

\item{reference_name}{Name of reference}

\item{treatment_name}{NAme of treatment}

\item{treatment_probability}{Probability of treatment affecting a person}

\item{reference_probability}{Probability of reference affecting a person}

\item{treatment_confidence_interval}{Confidence interval of treatment effect}
}
\value{
{ggplot2} object
}
\description{
Plot the axes, ticks and labels
}
