% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopViz.R
\name{.PlotConfidenceInterval}
\alias{.PlotConfidenceInterval}
\title{Plot the conifdence interval gradient.}
\usage{
.PlotConfidenceInterval(
  plot,
  treatment_probability,
  treatment_confidence_interval
)
}
\arguments{
\item{plot}{{ggplot2} object to which to add elements}

\item{treatment_probability}{Probability of treatment affecting a person}

\item{treatment_confidence_interval}{Confidence interval of treatment effect}
}
\value{
{ggplot2} object
}
\description{
Plot the conifdence interval gradient.
}
