% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopViz.R
\name{.PlotPartialGlyph}
\alias{.PlotPartialGlyph}
\title{Plot a partial person overlapping whole people.}
\usage{
.PlotPartialGlyph(
  plot,
  person_count,
  person_multiplier,
  glyph_spacing,
  glyph_colour,
  glyph_size,
  glyph_resolution,
  svg_text_raw
)
}
\arguments{
\item{plot}{{ggplot2} object to which to add elements}

\item{person_count}{total number of people represented in this category. Not the whole number represented in the graphic}

\item{person_multiplier}{The number of people that each glyph represents}

\item{glyph_spacing}{Spacing between glyphs}

\item{glyph_colour}{Colour to display the glyph}

\item{glyph_size}{Size of glyph}

\item{glyph_resolution}{Resolution of the glyph}

\item{svg_text_raw}{Raw text for SVG to be modified}
}
\value{
{ggplot2} object
}
\description{
Plot a partial person overlapping whole people.
}
