% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopViz.R
\name{.PlotTitle}
\alias{.PlotTitle}
\title{Add the title to the plot.}
\usage{
.PlotTitle(
  plot,
  title,
  person_count,
  outcome_name,
  reference_name,
  treatment_name,
  reference_probability,
  treatment_probability,
  population_name,
  treatment_plural
)
}
\arguments{
\item{plot}{{ggplot2} object to which to add elements}

\item{title}{Title to add to graphic. If NULL, the title will be created from the data. Defaults to NULL}

\item{person_count}{Number of people being represented}

\item{outcome_name}{Name of event being measured}

\item{reference_name}{Name of reference}

\item{treatment_name}{Name of treatment}

\item{reference_probability}{Probability of reference affecting a person}

\item{treatment_probability}{Probability of treatment affecting a person}

\item{population_name}{Name of the population}

\item{treatment_plural}{TRUE if the treatment is a plural}
}
\value{
{ggplot2} object
}
\description{
Add the title to the plot.
}
