% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_object_extraction.R
\name{ExtractMetafor}
\alias{ExtractMetafor}
\title{Title}
\usage{
ExtractMetafor(model)
}
\arguments{
\item{model}{}
}
\value{
List of plot options:
- "relative_effect": Point estimate of relative effect
- "confidence_interval": vector of lower and upper limits of confidence interval of relative effect
- "outcome_type": Type of outcome of analysis
}
\description{
Title
}
\examples{
ExtractMetafor(metafor_model)
}
