% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_object_extraction.R
\name{ExtractMetaStan}
\alias{ExtractMetaStan}
\title{Get PopViz data from {MetaStan} object (meta_stan)}
\usage{
ExtractMetaStan(model)
}
\arguments{
\item{model}{MetaStan analysis object from which to get PopViz data}
}
\value{
List of plot options:
- "relative_effect": Point estimate of relative effect
- "confidence_interval": vector of lower and upper limits of confidence interval of relative effect
- "outcome_type": Type of outcome of analysis
}
\description{
Get PopViz data from {MetaStan} object (meta_stan)
}
\examples{
ExtractMetaStan(metastan_model)
}
