% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_manipulation.R
\name{GetSvgText}
\alias{GetSvgText}
\title{Get text content from an SVG file, and modify with the given styling.}
\usage{
GetSvgText(
  filename = "svgs/person-super-narrow.svg",
  colour = "#000000",
  start_proportion = 0,
  end_proportion = 1
)
}
\arguments{
\item{filename}{Name of file from which to read text. Defaults to "person-solid.svg"}

\item{colour}{Hexidecimal colour to set on all paths within the svg. Defaults to "#000000"}

\item{start_proportion}{Proportion by which to truncate the left side of the image. Defaults to 0}

\item{end_proportion}{Proportion by which to include along the right side of the image. Defaults to 1}
}
\value{
Text extracted from the SVG file, and modified with the given styling
}
\description{
Get text content from an SVG file, and modify with the given styling.
}
\examples{
GetSvgText()
GetSvgText("cat.svg")
GetSvgText("cat.svg", 0.2, 0.8)
}
