% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifySvgText.R
\name{ModifySvgText}
\alias{ModifySvgText}
\title{Modify SVG text with the given styling.}
\usage{
ModifySvgText(
  svg_text,
  colour = "#000000",
  start_proportion = 0,
  end_proportion = 1
)
}
\arguments{
\item{svg_text}{Text content of a `.svg` file}

\item{colour}{Hexidecimal colour to set on all paths within the svg. Defaults to "#000000"}

\item{start_proportion}{Proportion by which to truncate the left side of the image. Defaults to 0}

\item{end_proportion}{Proportion by which to include along the right side of the image. Defaults to 1}
}
\value{
SVG text modified with the given styling
}
\description{
Modify SVG text with the given styling.
}
\examples{
svg_text <- GetSvgText()

ModifySvgText(svg_text)
ModifySvgText(svg_text, "#33aaff", 0.2, 0.8)
}
