% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopViz.R
\name{PopViz}
\alias{PopViz}
\title{Create a population visualisation for absolute effects.}
\usage{
PopViz(
  person_count,
  event_desired,
  outcome_name,
  reference_name,
  treatment_name,
  outcome_type,
  reference_probability,
  relative_effect,
  relative_confidence_interval,
  title = NULL
)
}
\arguments{
\item{person_count}{Number of people to display}

\item{event_desired}{TRUE if events are desired, else FALSE}

\item{outcome_name}{Name of event being measured}

\item{reference_name}{Name of reference intervention}

\item{treatment_name}{Name of intervention being compared}

\item{outcome_type}{Type of outcome being measured}

\item{reference_probability}{Probability of event in reference intervention}

\item{relative_effect}{Point estimate of relative effect}

\item{relative_confidence_interval}{Vector of lower and upper limits of confidence interval of relative effect}
}
\value{
ggplot2 plot object
}
\description{
Create a population visualisation for absolute effects.
}
\examples{
# For some reason, {ggplot2} must be loaded explicitly
library(ggplot2)

PopViz::PopViz(
  person_count = 5,
  event_desired = FALSE,
  outcome_name = "Adverse Events",
  reference_name = "Standard Care",
  treatment_name = "Treatment",
  outcome_type = "RD",
  reference_probability = 0.5,
  relative_effect = 0.3,
  relative_confidence_interval = c(0.1, 0.5)
)
}
