% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSvgText.R
\name{ReadSvgText}
\alias{ReadSvgText}
\title{Read text content from an SVG file.}
\usage{
ReadSvgText(
  filename = system.file("person-super-narrow.svg", package = "PopViz")
)
}
\value{
Text extracted from the SVG file
}
\description{
Read text content from an SVG file.
}
\examples{
GetSvgText()
GetSvgText("cat.svg")
}
