% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopViz.R
\name{.GetGlyphFile}
\alias{.GetGlyphFile}
\title{Get the glyph file based on the number of people on the plot.}
\usage{
.GetGlyphFile(person_count)
}
\arguments{
\item{person_count}{The number of people on the plot}
}
\value{
List containing:
- "filename": The path to the glyph file
- "size": the size of the glyph to be rendered
}
\description{
Get the glyph file based on the number of people on the plot.
}
