% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopViz.R
\name{.PlotTitle}
\alias{.PlotTitle}
\title{Add the title to the plot.}
\usage{
.PlotTitle(
  plot,
  title,
  person_count,
  outcome_name,
  reference_name,
  treatment_name,
  reference_probability,
  treatment_probability
)
}
\arguments{
\item{plot}{{ggplot2} object to which to add elements}

\item{title}{Title of the plot. Can be `NULL` if non specified}

\item{person_count}{Number of people being rendered}

\item{outcome_name}{Name of event being measured}

\item{reference_name}{Name of reference}

\item{treatment_name}{Name of treatment}

\item{reference_probability}{Probability of reference affecting a person}

\item{treatment_probability}{Probability of treatment affecting a person}
}
\value{
{ggplot2} object
}
\description{
Add the title to the plot.
}
