% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopViz.R
\name{PopViz}
\alias{PopViz}
\title{Create a population visualisation for absolute effects.}
\usage{
PopViz(
  NoPeople,
  DesireEvent,
  OutcomeName,
  TreatmentName,
  ComparatorName,
  OutcomeType,
  RelEff,
  RelConfInt,
  ComProb,
  ComConfInt,
  Title = NULL
)
}
\arguments{
\item{NoPeople}{Number of people to display}

\item{DesireEvent}{TRUE if events are desired, else FALSE}

\item{OutcomeName}{Name of event being measured}

\item{TreatmentName}{Name of intervention being compared}

\item{ComparatorName}{Name of reference intervention}

\item{OutcomeType}{Type of outcome being measured}

\item{RelEff}{Point estimate of relative effect}

\item{RelConfInt}{Vector of lower and upper limits of confidence interval of relative effect}

\item{ComProb}{Probability of event in reference intervention}
}
\value{
ggplot2 plot object
}
\description{
Create a population visualisation for absolute effects.
}
\examples{
library(ggplot2)

PopViz::PopViz(
  NoPeople = 50,
  DesireEvent = FALSE,
  OutcomeName = "Adverse Effects",
  TreatmentName = "Treatment",
  ComparatorName = "Standard Care",
  OutcomeType = "RD",
  ComProb = 0.5,
  RelEff = 0.2,
  RelConfInt = c(0.1, 0.3)
)
}
